fix: DealerTest에서 BetTest로 로직 이동
